Commands.md: Clarify BF.RESERVE's error_rate arg
The documentation related to BF.RESERVE's error_rate arg was a bit
misleading and confusing. Specifically, it said that greater values
would result in more space consumption when the opposite is true, and
described it as a "percentage" when it's a ratio, which might result in
people choosing values outside the correct (0, 1) range.