Fix crash when error rate is very high (#720)
Crash scenario:

```
BF.RESERVE b 0.99 3 NONSCALING
BF.ADD b 1
```

It seems like crash is only possible with NONSCALING filters with 
a very high error rate and relatively small capacity. Creating filter 
with these parameters does not make much sense but it can happen
if wrong arguments are passed. 

With the above parameters, we calculate number of required bits in 
the filter as zero and it leads to crash later. 

To fix the issue, we can limit minimum number of bits to 1.