RED-134847 support active defrag on all probabilistic data types. (#25)
The PR adds support for active defrag on all probabilistic data types. The implementation is strate forward, for each pointer in each datatype we call `RedisModule_DefragAlloc` to defrag the memory pointed by the poiter.

Notice that as part of the changes, t-digest-c library was modified to use RedisModule Allocator for memory allocation. This is not the case before this PR and it was wrong, the memory allocation should be done via Redis allocator.

Test was added to cover the new functionality.

(cherry picked from commit 00cc08d2d3284f025bc18b650d41301619abb606)